// src/payload-types.ts
export interface Calc {
  id: number;
  bez_zaloga: {
    annualInterestRate: number;
    minTermMonths: number;
    maxTermMonths: number;
    minLoanAmount: number;
    maxLoanAmount: number;
    termMonths: number;
    loanAmount: number;
  };
  pod_zalog: {
    annualInterestRate: number;
    annualInterestRateGold: number;
    minTermMonths: number;
    maxTermMonths: number;
    maxTermMonthsGold: number;
    minLoanAmount: number;
    maxLoanAmount: number;
    termMonths: number;
    loanAmount: number;
  };
  avtokredit: {
    annualInterestRate: number;
    minTermMonths: number;
    maxTermMonths: number;
    minLoanAmount: number;
    maxLoanAmount: number;
    downPaymentPercent: number;
    termMonths: number;
    loanAmount: number;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}